KUBIK-16 add type to API
