+++ packages/client/src/api/leaderBoardApi.ts
@@ -1,7 +1,18 @@
-import { BaseUrl } from '../utils/baseURL'
+import {BaseUrl} from '../utils/baseURL'
+
+export interface LeadersProps {
+  user_name: string
+  avatar: string
+  userId: number
+  score: number
+}
+
+export interface LeadersError {
+  reason: string;
+}
 
 export const LeaderBoardApi = {
-  async getAll() {
+  async getAll(): Promise<LeadersProps[] | LeadersError> {
     return fetch(`${BaseUrl}/leaderboard/all`, {
       body: '',
       method: 'POST',
