+++ packages/client/src/pages/LeaderBoardPage/LeaderBoardPage.tsx
@@ -1,37 +1,31 @@
-import React, { useEffect, useState } from 'react'
+import React, {useEffect, useState} from 'react'
 import GamerItem from '../../components/GamerItem/GamerItem'
-import { LeaderBoardApi } from '../../api/leaderBoardApi'
+import {LeaderBoardApi, LeadersError, LeadersProps} from '../../api/leaderBoardApi'
 import './LeaderBoardPage.scss'
 
-export interface LeadersProps {
-  user_name: string
-  avatar: string
-  userId: number
-  score: number
-}
-
 const LeaderBoardPage = () => {
-  const [leaders, setLeaders] = useState<LeadersProps[]>([])
+    const [leaders, setLeaders] = useState<LeadersProps[] | LeadersError>([]);
 
-  useEffect(() => {
-    getLeaders()
-  }, [])
+    useEffect(() => {
+        getLeaders()
+    }, [])
 
-  const getLeaders = () => {
-    LeaderBoardApi.getAll().then(response => {
-      if (!response.reason) {
-        setLeaders(response)
-      }
-    })
-  }
+    const getLeaders = (): void => {
+        const response: Promise<LeadersProps[] | LeadersError> = LeaderBoardApi.getAll();
+        response.then(response => {
+            if (Array.isArray(response)) {
+                setLeaders(response)
+            }
+        })
+    }
 
-  return (
-    <div className="lieder-board-page">
-      {leaders.map(leader => (
-        <GamerItem profile={leader} key={leader.userId} />
-      ))}
-    </div>
-  )
+    return (
+        <div className="lieder-board-page">
+            {leaders.map(leader => (
+                <GamerItem profile={leader} key={leader.userId}/>
+            ))}
+        </div>
+    )
 }
 
 export default LeaderBoardPage
